src/discord-bot/requirements.txt
aiosqlite==0.17.0
discord.py==1.7.3
python-dotenv==0.20.0
requests==2.27.1